Result.py: Add actions attribute
This adds a new attribute `actions`, a list of action objects,
to Result class. Default actions are added to this list by default.
This also changes the `init` and `from_values` method to support
bears to pass a list of bear specific actions while yielding a
Result object. This list of bear specific actions is added to
`actions` attribute.