Result.py: Add actions attribute
This adds support for bears to define their own custom actions. A new
`actions` attribute is added to Result class. It is a list of action
objects defined by the bears.
This also adds a new attribute `identity` to `get_metadata` function is
ResultAction module. This is used as a key for `action_dict`.
Also made changes in ConsoleInteraction and Processing module to add
actions to list of default actions.